Action working for click user event
The action is working for click user event. I was running into problems
because when I wasn't exporting the const SELECT_BLOG - - so on my
reducer, it wasn't able to pick up the string that I needed.  Unable to
read the action type, so it was still catching the console log, but was
having trouble with a specific blog.