Updating the components
Each component is up and working and connected a little style into it.
Trying to think about - - content, and google maps would be awesome, by
adding the longitute and latitude to each of the components, so I think
that's the next thing I might be looking into. It's close to mvp,
without styling, but - -  want to shrink the code, becasue I have a lot
of repitition and get my brain thinking more and more in terms of the
code.