Added markers
Added markers, added more content to the info window - - - added an icon
variable, which allowed me to change the icon size of what I was passing
through. Unsure if that is to the browser or the original file, but
either way.

Got rid of place and uniqueLocation - and changed them all to name.
Still trying to figure out how to style everything else.