Updates to Home Link.
I need to add this home link to the mobile site in some way - - and I
also - - need to figure out the transparency issues with a few of these
logos.

Also - need to write a paragraph - find a paragraph about each school
and implement it so there's simply more information.